+++ models/venue.js
@@ -3,8 +3,8 @@ module.exports = function(sequelize, DataTypes) {
   var venue = sequelize.define('venue', {
     name: DataTypes.STRING,
     url: DataTypes.STRING,
-    lat: DataTypes.INTEGER,
-    lng: DataTypes.INTEGER,
+    lat: DataTypes.FLOAT(15),
+    lng: DataTypes.FLOAT(15),
     address: DataTypes.STRING,
     city: DataTypes.STRING,
     state: DataTypes.STRING,
